feat: imporved testing on the server side
